Replace fixed (pixel) width sizes with Bootstrap classes
See http://getbootstrap.com/css/#grid

Note that the bootstrap 'grid' (rows and columns) makes use of media
queries in the way they describe, but that in the meantime the twelve
columns vary with the size of the screen— there's always twelve columns,
on a small screen they might be 50 pixels (50px) each, and on a
large screen they're 100px each.  A class of col-md-5 on a div or
something else acting as a column says that on small devices, stack this
div vertically with other divs, but on medium-sized devices (md for
medium) have this div or other block element take up five of the twelve
columns.